fix: move enums to separate file
This fixes them showing up as a suggested value, which is incorrect because they are only used for types.
